requirements.txt
alembic==1.7.1
click==7.1.2
ecdsa==0.17.0
Flask==1.1.4
Flask-Cors==3.0.10
Flask-Migrate==2.7.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
greenlet==1.1.1
gunicorn==20.1.0
importlib-resources==5.2.2
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.1.5
MarkupSafe==2.0.1
psycopg2-binary==2.9.1
pyasn1==0.4.8
python-dotenv==0.19.0
python-jose==3.3.0
rsa==4.7.2
six==1.16.0
SQLAlchemy==1.4.23
Werkzeug==1.0.1
zipp==3.5.0
